+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog]
 
 ## [Unreleased]
 
-- Introduce the beginnings of the Maths & Physics journey
+- Introduce the beginnings of the Maths & Physics journey, protected with basic
+  auth
 
 ## [Release 030] - 2019-11-12